# soroban-rpc: libpreflight: Remove base64 encoding between Go and Rust