--- a/cmd/soroban-rpc/lib/preflight.h
+++ b/cmd/soroban-rpc/lib/preflight.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 
-typedef struct CLedgerInfo {
+typedef struct ledger_info_t {
   uint32_t protocol_version;
   uint32_t sequence_number;
   uint64_t timestamp;
@@ -13,37 +13,47 @@ typedef struct CLedgerInfo {
   uint32_t min_persistent_entry_expiration;
   uint32_t max_entry_expiration;
   uint32_t auto_bump_ledgers;
-} CLedgerInfo;
+} ledger_info_t;
+
+typedef struct xdr_t {
+    unsigned char *xdr;
+    size_t        len;
+} xdr_t;
+
+typedef struct xdr_vector_t {
+    xdr_t  *array;
+    size_t len;
+} xdr_vector_t;
+
+typedef struct preflight_result_t {
+    char          *error; // Error string in case of error, otherwise null
+    xdr_vector_t  auth; // array of SorobanAuthorizationEntries
+    xdr_t         result; // XDR SCVal
+    xdr_t         transaction_data;
+    int64_t       min_fee; // Minimum recommended resource fee
+    xdr_vector_t  events; // array of XDR DiagnosticEvents
+    uint64_t      cpu_instructions;
+    uint64_t      memory_bytes;
+    xdr_t         pre_restore_transaction_data; // SorobanTransactionData XDR for a prerequired RestoreFootprint operation
+    int64_t       pre_restore_min_fee; // Minimum recommended resource fee for a prerequired RestoreFootprint operation
+} preflight_result_t;
+
+preflight_result_t *preflight_invoke_hf_op(uintptr_t handle, // Go Handle to forward to SnapshotSourceGet
+                                           uint64_t bucket_list_size, // Bucket list size of current ledger
+                                           const xdr_t invoke_hf_op, // InvokeHostFunctionOp XDR
+                                           const xdr_t source_account, // AccountId XDR
+                                           const ledger_info_t ledger_info);
+
+preflight_result_t *preflight_footprint_expiration_op(uintptr_t   handle, // Go Handle to forward to SnapshotSourceGet
+                                                      uint64_t bucket_list_size, // Bucket list size of current ledger
+                                                      const xdr_t op_body, // OperationBody XDR
+                                                      const xdr_t footprint, // LedgerFootprint XDR
+                                                      uint32_t    current_ledger_seq); // Current ledger sequence
 
 
-typedef struct CPreflightResult {
-    char *error; // Error string in case of error, otherwise null
-    char **auth; // NULL terminated array of XDR SorobanAuthorizationEntrys in base64
-    char *result; // XDR SCVal in base64
-    char *transaction_data; // SorobanTransactionData XDR in base64
-    int64_t min_fee; // Minimum recommended resource fee
-    char **events; // NULL terminated array of XDR DiagnosticEvents in base64
-    uint64_t cpu_instructions;
-    uint64_t memory_bytes;
-    char *pre_restore_transaction_data; // SorobanTransactionData XDR in base64 for a prerequired RestoreFootprint operation
-    int64_t pre_restore_min_fee; // Minimum recommended resource fee for a prerequired RestoreFootprint operation
-} CPreflightResult;
+// LedgerKey XDR to LedgerEntry XDR
+extern xdr_t SnapshotSourceGet(uintptr_t handle, xdr_t ledger_key, int include_expired);
 
-CPreflightResult *preflight_invoke_hf_op(uintptr_t handle, // Go Handle to forward to SnapshotSourceGet and SnapshotSourceHas
-                                         uint64_t bucket_list_size, // Bucket list size of current ledger
-                                         const char *invoke_hf_op, // InvokeHostFunctionOp XDR in base64
-                                         const char *source_account, // AccountId XDR in base64
-                                         const struct CLedgerInfo ledger_info);
+void free_preflight_result(preflight_result_t *result);
 
-CPreflightResult *preflight_footprint_expiration_op(uintptr_t handle, // Go Handle to forward to SnapshotSourceGet and SnapshotSourceHas
-                                                    uint64_t bucket_list_size, // Bucket list size of current ledger
-                                                    const char *op_body, // OperationBody XDR in base64
-                                                    const char *footprint, // LedgerFootprint XDR in base64
-                                                    uint32_t current_ledger_seq); // Current ledger sequence
-
-// LedgerKey XDR in base64 string to LedgerEntry XDR in base64 string
-extern char *SnapshotSourceGet(uintptr_t handle, char *ledger_key, int include_expired);
-
-void free_preflight_result(CPreflightResult *result);
-
-extern void FreeGoCString(char *str);
+extern void FreeGoXDR(xdr_t xdr);